cotask/requirements/prod.txt
Django==5.2.2
psycopg2-binary==2.9.10
python-dotenv==1.1.0
djangorestframework==3.16.0
django-modeltranslation==0.19.15
Pillow==11.2.1
gunicorn
channels
channels_redis
daphne